Moved and renamed job usecases
